+++ src/Mapping/GpApiMapping.php
@@ -55,6 +55,7 @@ use GlobalPayments\Api\Entities\Reporting\PayByLinkSummary;
 use GlobalPayments\Api\Entities\Reporting\StoredPaymentMethodSummary;
 use GlobalPayments\Api\Entities\Reporting\TransactionSummary;
 use GlobalPayments\Api\Entities\RiskAssessment;
+use GlobalPayments\Api\Entities\Installment;
 use GlobalPayments\Api\Entities\ThirdPartyResponse;
 use GlobalPayments\Api\Entities\ThreeDSecure;
 use GlobalPayments\Api\Entities\Transaction;
@@ -89,6 +90,8 @@ class GpApiMapping
 
     const BATCH_CLOSE = 'CLOSE';
 
+    /** @var Installment */
+     public $installment;
     /**
      * Map a response to a Transaction object for further chaining
      *
@@ -177,6 +180,11 @@ class GpApiMapping
         if (!empty($response->payer)) {
             $transaction->payerDetails = self::mapPayerDetails($response->payer);
         }
+
+        if (!empty($response->installment)) {
+            $transaction->installment = $response->installment;
+        }
+    
         return $transaction;
     }